requirements.txt
azure-identity
azure-mgmt-containerinstance